Rename manifest handler
