+++ packages/cli/src/cmds/manifest/manifestHandler.ts
